refactor(PositionConverterUtil): change class name and separate private method logic to inner class
