+++ src/main/java/chess2/util2/PieceGeneratorUtil.java
@@ -5,7 +5,7 @@ import static chess2.domain2.board2.piece2.PieceType.KING;
 import static chess2.domain2.board2.piece2.PieceType.KNIGHT;
 import static chess2.domain2.board2.piece2.PieceType.QUEEN;
 import static chess2.domain2.board2.piece2.PieceType.ROOK;
-import static chess2.util2.PositionUtil.FILES_TOTAL_SIZE;
+import static chess2.util2.PositionConverterUtil.FILES_TOTAL_SIZE;
 
 import chess2.domain2.board2.Position;
 import chess2.domain2.board2.piece2.Color;
